Moved search response parser to registry-common project
