+++ src/main/java/gov/nasa/pds/registry/mgr/dao/DataExporter.java
@@ -9,9 +9,9 @@ import org.elasticsearch.client.RestClient;
 
 import gov.nasa.pds.registry.common.es.client.EsClientFactory;
 import gov.nasa.pds.registry.common.es.client.EsUtils;
+import gov.nasa.pds.registry.common.es.client.SearchResponseParser;
 import gov.nasa.pds.registry.mgr.util.CloseUtils;
 import gov.nasa.pds.registry.mgr.util.es.EsDocWriter;
-import gov.nasa.pds.registry.mgr.util.es.SearchResponseParser;
 
 
 public abstract class DataExporter
